Fix definition of collections.namedtuple.
Also includes the following changes:

Accept str and bytes as unicode, and int as a float. 

Side effect: drop duplicate types in Union[]

This expands existing code in abstract.Class.match_instance_against_type that allowed us to accept None as a bool.

Motivation is to remove false positives coming from typeshed, but type-matching may be tightened back up in the future.

Change happens in two places, long-term it may be possible to have type-comparison logic in just one.

--
MOS_MIGRATED_REVID=132577949

+++ pytype/tests/test_pyi.py
@@ -47,7 +47,7 @@ class PYITest(test_inference.InferenceTest):
     with utils.Tempdir() as d:
       d.create_file("mod.pyi", """
         from typing import Optional, List, Any, IO
-        def split(s: Optional[float]) -> List[str, ...]: ...
+        def split(s: Optional[int]) -> List[str, ...]: ...
       """)
       ty = self.Infer("""\
         import mod
@@ -56,7 +56,7 @@ class PYITest(test_inference.InferenceTest):
       """, deep=True, solve_unknowns=True, pythonpath=[d.path])
       self.assertTypesMatchPytd(ty, """
         mod = ...  # type: module
-        def g(x: NoneType or float) -> List[str, ...]
+        def g(x: NoneType or int) -> List[str, ...]
       """)
 
   def testClasses(self):
